requirements.txt
dify_plugin>=0.4.0,<0.7.0
openai>=1.0.0,<2.0.0
google-genai>=1.0.0
